requirements.txt
pandas
openpyxl
pymupdf
tqdm
python-dotenv
scikit-learn
rank-bm25
sentence-transformers
faiss-cpu
jupyter
matplotlib
google-genai